Move LightwalletdTestType to the lightwalletd module
